+++ src/main/java/domain/score/Referee.java
@@ -1,11 +1,9 @@
-package domain.game;
+package domain.score;
 
 import domain.card.Cards;
 import domain.card.DealerCards;
 import domain.card.PlayerCards;
 import domain.player.Name;
-import domain.score.Outcome;
-import domain.score.ScoreBoard;
 
 import java.util.Map;
