+++ src/main/java/domain/score/ScoreBoard.java
@@ -1,6 +1,6 @@
 package domain.score;
 
-import domain.player.Bet;
+import domain.game.Bet;
 import domain.player.Name;
 
 import java.util.*;
