+++ src/test/java/domain/game/RefereeTest.java
@@ -4,8 +4,8 @@ import domain.card.Card;
 import domain.card.DealerCards;
 import domain.card.PlayerCards;
 import domain.card.Shape;
-import domain.player.Bet;
 import domain.player.Name;
+import domain.score.Referee;
 import domain.score.Revenue;
 import domain.score.ScoreBoard;
 import org.junit.jupiter.api.BeforeEach;
