+++ src/test/java/domain/player/BetTest.java
@@ -1,5 +1,6 @@
 package domain.player;
 
+import domain.game.Bet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
